Moved some program services to keep code readable. Added comments to some code.
