+++ CitationWebAPI/CitationWebAPI.csproj
@@ -20,6 +20,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Interfaces\Repository\" />
     <Folder Include="Migrations" />
   </ItemGroup>
