+++ CitationWebAPI/Models/User.cs
@@ -1,5 +1,10 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
+/**
+* Defines properties for a user that will be stored in database
+* Used by Entity Framework Core
+*/
+
 namespace CitationWebAPI.Models
 {
     public class User
