+++ CitationWebAPI/Program.cs
@@ -1,7 +1,4 @@
-using CitationWebAPI.Data;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.OpenApi.Any;
-using System.Text.RegularExpressions;
+using CitationWebAPI;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -13,46 +10,9 @@ builder.Services
     .AddControllers(options => options.UseDateOnlyTimeOnlyStringConverters())
     .AddJsonOptions(options => options.UseDateOnlyTimeOnlyStringConverters());
 
-builder.Services.AddSwaggerGen(options =>
-    options.MapType<DateOnly>(() => new Microsoft.OpenApi.Models.OpenApiSchema
-    {
-        Type = "string",
-        Format = "date",
-        Example = new OpenApiString("2022-10-05")
-    })
-);
-
-builder.Services.AddSwaggerGen(options =>
-    options.MapType<TimeOnly>(() => new Microsoft.OpenApi.Models.OpenApiSchema
-    {
-        Type = "string",
-        Format = "time",
-        Example = new OpenApiString("23:15:00")
-    })
-);
-
-
-builder.Services.AddDbContext<DataContext>(options =>
-{
-    if (Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT") == "Production")
-    {
-        var m = Regex.Match(Environment.GetEnvironmentVariable("DATABASE_URL")!, @"postgres://(.*):(.*)@(.*):(.*)/(.*)");
-        options.UseNpgsql($"Server={m.Groups[3]};Port={m.Groups[4]};User Id={m.Groups[1]};Password={m.Groups[2]};Database={m.Groups[5]};sslmode=Prefer;Trust Server Certificate=true");
-    }
-    else // In Development Environment
-    {
-        // So, use a local Connection
-        options.UseNpgsql(builder.Configuration.GetConnectionString("LocalConnection"));
-    }
-});
-
-// Use same origin as angular app (can add more later)
-builder.Services.AddCors(opt => opt.AddPolicy(
-    name: "CitationOrigins", policy => policy
-    .WithOrigins("http://localhost:4200", "http://localhost:8080", "https://traffic-citation-frontend.herokuapp.com")
-    .AllowAnyMethod()
-    .AllowAnyHeader())); 
-    
+builder.Services.ConfigureCors();
+builder.Services.ConfigureDbContext(builder);
+builder.Services.AddSwaggerGenTypes();
 
 var app = builder.Build();
